# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  user: "user",
  demo: "demo",
  support: "support",
  data_entry: "data_entry",
  property_manager: "property_manager",
  admin: "admin",
  super_admin: "super_admin",
} as const;

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  /** @nullable */
  phone: string | null;
  isAdmin: boolean;
  role: AuthUserRole;
  isDisabled: boolean;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface UsersListEnvelope {
  users: AuthUser[];
}

export interface UpdateMyProfileRequest {
  /**
   * @maxLength 100
   * @nullable
   */
  firstName?: string | null;
  /**
   * @maxLength 100
   * @nullable
   */
  lastName?: string | null;
  /**
   * @maxLength 1000
   * @nullable
   */
  profileImageUrl?: string | null;
  /**
   * @maxLength 30
   * @nullable
   */
  phone?: string | null;
}

export interface UpdateUserRequest {
  /**
   * @maxLength 100
   * @nullable
   */
  firstName?: string | null;
  /**
   * @maxLength 100
   * @nullable
   */
  lastName?: string | null;
  /**
   * @maxLength 255
   * @nullable
   */
  email?: string | null;
  /**
   * @maxLength 1000
   * @nullable
   */
  profileImageUrl?: string | null;
  /**
   * @maxLength 30
   * @nullable
   */
  phone?: string | null;
  isAdmin?: boolean;
  isDisabled?: boolean;
  /**
   * @minLength 8
   * @maxLength 200
   */
  password?: string;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};