# lib/api-zod/src/generated/types/updateUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateUserRequest {
  /**
   * @maxLength 100
   * @nullable
   */
  firstName?: string | null;
  /**
   * @maxLength 100
   * @nullable
   */
  lastName?: string | null;
  /**
   * @maxLength 255
   * @nullable
   */
  email?: string | null;
  /**
   * @maxLength 1000
   * @nullable
   */
  profileImageUrl?: string | null;
  /**
   * @maxLength 30
   * @nullable
   */
  phone?: string | null;
  isAdmin?: boolean;
  isDisabled?: boolean;
  /**
   * @minLength 8
   * @maxLength 200
   */
  password?: string;
}